refactor(ensnode-sdk): move relevant types from `shared` lib to `ens` lib
